Copied 2 files from Abel and corrected all 7 mistakes made, realised I shouldn't apply my own logic to the slides. Also can't update badge yet either

+++ MyCheeseShop/Model/CartItem.cs
@@ -1,6 +1,4 @@
 ﻿namespace MyCheeseShop.Model
-
-
 {
     public class CartItem
     {
@@ -8,7 +6,6 @@
 
         public int Quantity { get; set; }
 
-        public decimal Total => Cheese.Price * Quantity;
-
+        public decimal Total => (decimal)(Cheese.Price * Quantity);
     }
 }
